Decompose BackupService, and fix two CodeQL findings
Closes #1092.

## The split

backend/src/backup/backup.service.ts was ~2,600 lines, three times the
repository ceiling, and combined four independently testable concerns
behind one set of private methods. Cancellation, transaction lifetime,
object-store side effects and schema restore ordering all shared one file
and one private state, which is what made the security and lifecycle
review of that path hard.

It is now a facade -- one delegation per method, no DataSource, no storage
provider -- over four components:

- BackupExportService: the REPEATABLE READ snapshot, the streamed and
  buffered assemblies, the attachment bytes an object store contributes,
  the completeness report.
- BackupRestoreService: the processing gate, decryption, decompression,
  format validation, re-authentication ordering, id remapping, and the one
  transaction the rest runs inside.
- BackupAttachmentTransferService: carried bytes, the legacy ownership
  proof, staging, and both cleanup paths. The only file in the module that
  touches the object store, which is what makes the ordering rule -- write
  before the commit, delete after it -- reviewable at all.
- BackupRestoreDatabaseService: teardown, currency preparation, row
  inserts, deferred-FK repair. Every method takes the caller's
  EntityManager and opens no transaction of its own.

Plus backup-format.ts (the file format all four share),
export-table-queries.ts (what is read, in what order, and what is
deliberately not), attachment-integrity.util.ts (one bytes-vs-metadata
comparison, three provenances) and backup-password.util.ts.

No behaviour change: all 105 SQL statements moved byte-identically, and
the public surface the controller, the auto-backup cron and the support
export use is unchanged. Every source file in the module is now under the
800-line ceiling.

module-shape.spec.ts holds the result: a line-count ceiling with a
shrink-only grandfather list, and a scan that fails if the facade
re-acquires a DataSource or a storage provider -- the dependency is
cheaper to check than the line count and fails earlier.

## A split silently disarms a source-scanning guard

Four guards named backup.service.ts: the two attachment readers, the
trigger-DDL ban, the export's bytea encoding, and the "one predicate"
check in currency-references.spec.ts. Every one would have gone on passing
while scanning a file the code had moved out of. They now walk the module
directory, or throw when their marker is missing rather than returning an
empty match set. The rule is written down in backend/CLAUDE.md.

## CodeQL

js/polynomial-redos, backup.controller.ts: the base64 round-trip check
stripped padding with /=+$/ applied to an attacker-supplied, as yet
unauthenticated header. On "=".repeat(n) + "a" the engine restarts the
scan at every position, so the cost is quadratic and paid on the event
loop -- one request stalls every other one in the process. Replaced with a
linear backwards walk. The regression test gives 200k characters a
two-second budget: fixed it takes milliseconds, and the old expression
took 177 seconds when it was put back to check.

js/biased-cryptographic-random, support-backup.service.ts: the pseudonym
allocator took `randomBytes(1)[0] % 26`. 256 is not a multiple of 26, so
A-V came up 10/256 each and W-Z 9/256 -- an 11% excess across twenty-two
letters. Now rejection-sampled above the largest multiple of 26 that fits
in a byte. Two tests: a 260k-draw distribution check (the biased path sits
~18 sigma outside the tolerance), and a scan of src/ that fails on any
`%` applied to a crypto random draw outside the one function that rejects
the tail first.
